Do not scale inputs with a standard deviation of zero.
If the scaling parameters are set via a training dataset and for a given
input/output, all patterns have the same value, this input/output's
standard deviation will be zero. Scaling this input/output then divides
by zero.

This adds a check in the scale/descale functions that makes libfann
forego all inputs and outputs whose standard deviation is zero.